requirements.txt


# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Firebase Admin SDK
firebase-admin==6.2.0

# LangChain and AI dependencies
langchain==0.0.350
langchain-openai==1.3.7

langchain-mcp-adapters==0.0.1
langgraph==0.0.20
langgraph-supervisor==0.0.1

# Agentic RAG System Dependencies
pydantic-ai==0.0.13
openai==1.3.7

# Anthropic for MCP integration
anthropic==0.8.1

# Vector Database and Embeddings
asyncpg==0.29.0
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
pinecone-client==2.2.4

# Knowledge Graph (Neo4j)
neo4j==5.15.0

# ML and NLP
torch==2.1.1
transformers==4.36.0
numpy==1.24.3
scikit-learn==1.3.2

# Image processing for YouTube verification
Pillow==10.1.0

# Google Cloud Vision (for production image recognition)
google-cloud-vision==3.4.4

# Data processing and utilities
pydantic==2.5.0
pydantic-settings==2.1.0

# Analytics Platform Dependencies
pandas==2.1.4
openpyxl==3.1.2
pyarrow==14.0.1
aiohttp==3.9.1

# HTTP client for API calls
httpx==0.25.2
aiohttp==3.9.1

# Async support
asyncio-mqtt==0.16.1

# Logging and monitoring
structlog==23.2.0

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database and caching
redis==5.0.1
motor==3.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production deployment
gunicorn==21.2.0 